% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cuminc.R
\name{extract_cuminc}
\alias{extract_cuminc}
\title{Extract data from cuminc object}
\usage{
extract_cuminc(cuminc_obj)
}
\arguments{
\item{cuminc_obj}{an object returned from the cmprsk::cuminc() function}
}
\value{
a data frame containing the cumulative incidence estimates
}
\description{
This function extracts the individual estimate information from a cuminc
object. This is mainly useful when plotting the cumulative incidence
function
}
\examples{

df <-
  data.frame(
    fail_time = 12 * rexp(100),
    group = paste0('group_', sample(c('a','b','c'), 100, replace = TRUE)),
    fail_status = sample(0:2, 100, replace = TRUE)
  )

extract_cuminc(
  cmprsk::cuminc(
      df$fail_time,
      df$fail_status,
      df$group
    )
)

}
\author{
Patrick Hilden
}
