% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_survfit.R
\name{extract_survfit}
\alias{extract_survfit}
\title{Extract data from survfit object}
\usage{
extract_survfit(survfit_obj)
}
\arguments{
\item{survfit_obj}{an object returned from the survival::survfit() function}
}
\value{
a data frame containing the estimate of the survival function
}
\description{
This function extracts the individual estimate information from a survfit
object. This is mainly useful when plotting the survival function
}
\examples{

df <- data.frame(
  fail_time = 12 * rexp(100),
  group = paste0('group_', sample(1:3, 100, replace = TRUE)),
  fail_status = sample(0:1, 100, replace = TRUE)
)

extract_survfit(
  survival::survfit(
    survival::Surv(fail_time, fail_status) ~ group,
    data = df
  )
)

}
\author{
Patrick Hilden
}
