% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recent.R
\name{load_recent}
\alias{load_recent}
\title{load the most recent .rdata file in a given directory}
\usage{
load_recent(dir_path)
}
\arguments{
\item{dir_path}{path to the directory containing .rdata files}
}
\description{
I typically have multiple .rdata files for a given project due to updates in the data over time. Each of these .rdata files is dated and this function will load the most recent (via sorting)
}
